backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.11
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
pandas>=2.0.0
pyarrow>=12.0.0
python-multipart>=0.0.6
rank_bm25>=0.2.2
numpy>=1.24.0
pydantic-settings>=2.0.0
requests>=2.31.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
